test(did): split the test vector reading code in a separate file/package
tests are much cleaner and explicit now
